created main screen of the app, PlotSurfaceScreen.
